+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Fix a bug where experiment flags passed via command-line were ignored.
 * Support specifying experimental features in `dartdoc_options.yaml`.
 * Document using the public name for private named parameters.
+* Deprecate the `--use-categories` option.
 
 ## 9.0.3
 * Require analyzer 10.1.0 APIs.